requirements.txt
pytest==6.0.1
coverage==5.2.1
pylint==2.5.3
Flask==1.1.2
Flask-Cors==3.0.9
hypothesis==5.24.3
requests==2.24.0
virtualenv==20.0.31
pylint-venv==2.1.1
mypy==0.782
PyJWT==1.7.1
attrs==20.2.0
